minor refactor and edit of default value in notifications of is approved to true so only false events should be approved
