+++ backend/stores/event_users_store.py
@@ -32,9 +32,8 @@ class EventUsersStore:
             raise Exception(f"Failed to get workers by event: {e}")
 
     @staticmethod
-    def update_record(event_id: int, worker_id: int, update_data: dict):
-        record = EventUsers.query.filter_by(event_id=event_id, worker_id=worker_id).first()
-        if record:
-            record.update(update_data)
-            return record
+    def update_event_users(event_id: int, worker_id: int, update_data: dict):
+        if event_users := EventUsers.query.filter_by(event_id=event_id, worker_id=worker_id).first():
+            event_users.update(update_data)
+            return event_users
         return None
