+++ backend/utils/openai_utils.py
@@ -1,8 +1,9 @@
 import openai
-from SECRETS import OPEN_AI_KEY
+from backend.SECRETS import OPEN_AI_KEY
 
-# Set the OpenAI API key
+
 openai.api_key = OPEN_AI_KEY
+
 
 def generate_event_description(prompt: str) -> str:
     """
